refactoring and performance updates
