+++ queries.py
@@ -7,31 +7,32 @@ from shapely import wkb
 import streamlit as st
 
 import credentials
+from constants import STATES
 
-fred_tables = [
+FRED_TABLES = [
     'burdened_households',
     # 'homeownership_rate',
     'income_inequality',
     'population_below_poverty',
     'single_parent_households',
-    # 'snap_benefits_recipients',
+    'snap_benefits_recipients',
     'unemployment_rate',
     'resident_population',
 ]
 
-static_tables = [
+STATIC_TABLES = [
     'chmura_economic_vulnerability_index',
     'median_rents',
     'fair_market_rents'
 ]
 
-static_columns = {
+STATIC_COLUMNS = {
     'chmura_economic_vulnerability_index': ['VulnerabilityIndex', 'Rank'],
-    'fair_market_rents': ['fmr_0', 'fmr_1', 'fmr_2', 'fmr_3', 'fmr_4', ],
-    'median_rents': ['rent50_0', 'rent50_1', 'rent50_2', 'rent50_3', 'rent50_4', ]
+    'fair_market_rents': ['fmr_0', 'fmr_1', 'fmr_2', 'fmr_3', 'fmr_4'],
+    'median_rents': ['rent50_0', 'rent50_1', 'rent50_2', 'rent50_3', 'rent50_4']
 }
 
-table_headers = {
+TABLE_HEADERS = {
     'burdened_households': 'Burdened Households',
     'homeownership_rate': 'Home Ownership',
     'income_inequality': 'Income Inequality',
@@ -42,7 +43,7 @@ table_headers = {
     'resident_population': 'Resident Population',
 }
 
-table_units = {
+TABLE_UNITS = {
     'burdened_households': '%',
     'homeownership_rate': '%',
     'income_inequality': 'Ratio',
@@ -119,19 +120,17 @@ def table_names_query() -> list:
     return res
 
 
-def latest_data_census_tracts(state: str, counties, tables) -> pd.DataFrame:
+@st.cache(ttl=3600)
+def latest_data_census_tracts(state: str, counties: list, tables: list) -> pd.DataFrame:
     conn, engine = init_connection()
     cur = conn.cursor()
     tracts_df = census_tracts_geom_query(counties, state)
-    if len(counties) > 1:
-        # where_clause = 'WHERE id_index.state_name = ' + "'" + state + "'" + ' ' + 'AND id_index.county_name IN ' + str(tuple(counties))
-        where_clause = 'WHERE id_index.state_name = ' + "'" + state + "'" + ' ' + 'AND id_index.county_name IN ' + str(
-            tuple(counties))
-    if len(counties) == 1:
-        # where_clause = 'WHERE id_index.state_name = ' + "'" + state + "'" + ' ' + 'AND id_index.county_name IN (' + "'" + counties[0] + "'" + ')'
-        where_clause = f"WHERE id_index.state_name ='{state}' AND id_index.county_name IN ('{counties[0]}')"
+    counties_str = str(tuple(counties)).replace(',)', ')')
+    where_clause = f"WHERE id_index.state_name ='{state}' AND id_index.county_name IN {counties_str}"
+
     for table_name in tables:
-        cur.execute(f"""SELECT {table_name}.*, id_index.county_name, id_index.county_id, id_index.state_name, resident_population_census_tract.tot_population_census_2010
+        cur.execute(f"""SELECT {table_name}.*, id_index.county_name, id_index.county_id, id_index.state_name, 
+        resident_population_census_tract.tot_population_census_2010
             FROM {table_name} 
             INNER JOIN id_index ON {table_name}.tract_id = id_index.tract_id
             INNER JOIN resident_population_census_tract ON {table_name}.tract_id = resident_population_census_tract.tract_id
@@ -145,6 +144,21 @@ def latest_data_census_tracts(state: str, counties, tables) -> pd.DataFrame:
         tracts_df = tracts_df.loc[:, ~tracts_df.columns.duplicated()]
 
     return tracts_df
+
+
+def load_distributions() -> tuple:
+    metro_areas = generic_select_query('housing_stock_distribution', [
+        'location',
+        '0_br_pct',
+        '1_br_pct',
+        '2_br_pct',
+        '3_br_pct',
+        '4_br_pct'
+    ])
+    locations = list(metro_areas['location'])
+    metro_areas.set_index('location', inplace=True)
+
+    return metro_areas, locations
 
 
 def policy_query() -> pd.DataFrame:
@@ -168,8 +182,8 @@ def latest_data_single_table(table_name: str, require_counties: bool = True) -> 
         'SELECT DISTINCT ON (county_id) '
         'county_id, date AS "{} Date", value AS "{} ({})" '
         'FROM {} '
-        'ORDER BY county_id , "date" DESC'.format(table_headers[table_name], table_headers[table_name],
-                                                  table_units[table_name], table_name))
+        'ORDER BY county_id , "date" DESC'.format(TABLE_HEADERS[table_name], TABLE_HEADERS[table_name],
+                                                  TABLE_UNITS[table_name], table_name))
     results = cur.fetchall()
     colnames = [desc[0] for desc in cur.description]
     conn.close()
@@ -183,7 +197,7 @@ def latest_data_single_table(table_name: str, require_counties: bool = True) -> 
 
 def latest_data_all_tables() -> pd.DataFrame:
     counties_df = counties_query()
-    for table_name in fred_tables:
+    for table_name in FRED_TABLES:
         table_output = latest_data_single_table(table_name, require_counties=False)
         counties_df = counties_df.merge(table_output)
     chmura_df = static_data_single_table('chmura_economic_vulnerability_index', ['VulnerabilityIndex'])
@@ -296,32 +310,24 @@ def census_tracts_geom_query(counties, state) -> pd.DataFrame:
     return geom_df
 
 
-def list_tables():
-    conn, engine = init_connection()
-
-    conn.close()
-    return
-
-
 def static_data_all_table() -> pd.DataFrame:
     counties_df = counties_query()
-    for table_name in static_tables:
-        table_output = static_data_single_table(table_name, static_columns[table_name])
+    for table_name in STATIC_TABLES:
+        table_output = static_data_single_table(table_name, STATIC_COLUMNS[table_name])
         counties_df = counties_df.merge(table_output)
     return counties_df
 
 
 def output_data(df: pd.DataFrame, table_name: str = 'fred_tables', ext: str = 'xlsx') -> str:
-    if not os.path.isdir('Output'):
-        os.mkdir('Output')
+    path = f'Output/{table_name}.{ext}'
     if ext == 'pk':
-        path = 'Output/{}.pk'.format(table_name)
         df.to_pickle(path)
     elif ext == 'xlsx':
-        path = 'Output/{}.xlsx'.format(table_name)
         df.to_excel(path)
+    elif ext == 'csv':
+        df.to_csv(path)
     else:
-        print('Only .pk and .xlsx outputs are currently supported.')
+        print('Only .pk, .csv, and .xlsx outputs are currently supported.')
         sys.exit()
     return path
 
@@ -329,14 +335,103 @@ def output_data(df: pd.DataFrame, table_name: str = 'fred_tables', ext: str = 'x
 def fmr_data():
     conn, engine = init_connection()
     cur = conn.cursor()
-    cur.execute(
-        'SELECT state_full as "State", countyname as "County Name" '
-        'FROM fair_market_rents'
-    )
+    cur.execute('SELECT state_full as "State", countyname as "County Name" FROM fair_market_rents;')
     colnames = [desc[0] for desc in cur.description]
     results = cur.fetchall()
     conn.close()
     return pd.DataFrame(results, columns=colnames)
+
+
+def filter_state(data: pd.DataFrame, state: str) -> pd.DataFrame:
+    return data[data['State'].str.lower() == state.lower()]
+
+
+def filter_counties(data: pd.DataFrame, counties: list) -> pd.DataFrame:
+    counties = [_.lower() for _ in counties]
+    return data[data['County Name'].str.lower().isin(counties)]
+
+
+def load_all_data() -> pd.DataFrame:
+    if os.path.exists("Output/all_tables.xlsx"):
+        try:
+            res = input('Previous data found. Use data from local `all_tables.xlsx`? [y/N]')
+            if res.lower() == 'y' or res.lower() == 'yes':
+                df = pd.read_excel('Output/all_tables.xlsx')
+            else:
+                df = latest_data_all_tables()
+        except:
+            print('Something went wrong with the Excel file. Falling back to database query.')
+            df = latest_data_all_tables()
+    else:
+        df = latest_data_all_tables()
+
+    return df
+
+
+def clean_data(data: pd.DataFrame) -> pd.DataFrame:
+    data.set_index(['State', 'County Name'], drop=True, inplace=True)
+
+    data.drop([
+        'Burdened Households Date',
+        'Income Inequality Date',
+        'Population Below Poverty Line Date',
+        'Single Parent Households Date',
+        'Unemployment Rate Date',
+        'Resident Population Date',
+        'SNAP Benefits Recipients Date'
+    ], axis=1, inplace=True)
+
+    data.rename({'Vulnerability Index': 'COVID Vulnerability Index'}, axis=1, inplace=True)
+
+    data = data.loc[:, ~data.columns.str.contains('^Unnamed')]
+
+    return data
+
+
+def get_existing_policies(df: pd.DataFrame) -> pd.DataFrame:
+    policy_df = policy_query()
+    temp_df = df.merge(policy_df, on='county_id')
+    if not temp_df.empty and len(df) == len(temp_df):
+        if st._is_running_with_streamlit:
+            if st.checkbox('Use existing policy data?'):
+                return temp_df
+        else:
+            res = input('Policy data found in database. Use this data? [Y/n]').strip()
+            if res.lower() == 'y' or res.lower() == 'yes' or res == '':
+                return temp_df
+
+    else:
+        policy_df = pd.read_excel('Policy Workbook.xlsx', sheet_name='Analysis Data')
+        temp_df = df.merge(policy_df, on='County Name')
+        if not temp_df.empty and len(df) == len(temp_df):
+            return temp_df
+        else:
+            print(
+                "INFO: Policy data not found. Check that you've properly filled in the Analysis Data page in `Policy Workbook.xlsx` with the counties you're analyzing.")
+
+    return df
+
+
+@st.cache(ttl=3600, allow_output_mutation=True)
+def get_county_data(state: str, counties: list = None, policy: bool = False):
+    df = load_all_data()
+    df = filter_state(df, state)
+    if counties is not None:
+        df = filter_counties(df, counties)
+    if policy:
+        df = get_existing_policies(df)
+    df = clean_data(df)
+    return df
+
+
+@st.cache(ttl=3600)
+def get_national_county_data():
+    frames = []
+    for s in STATES:
+        tmp_df = get_county_data(s)
+        frames.append(tmp_df)
+    df = pd.concat(frames)
+    return df
 
 
 if __name__ == '__main__':
