+++ utils.py
@@ -3,6 +3,7 @@ import pandas as pd
 from six import BytesIO
 import geopandas as gpd
 
+
 def to_excel(df: pd.DataFrame):
     output = BytesIO()
     writer = pd.ExcelWriter(output, engine='xlsxwriter')
